Add AutoMapper and crud controller

+++ Entities/BaseEntity.cs
@@ -18,7 +18,7 @@ namespace Entities
         public TKey Id { get; set; }
     }
 
-    public abstract class BaseEntity : BaseEntity<int>
+    public abstract class BaseEntity : BaseEntity<long>
     {
     }
